Consolidate to single Blazor Server project and remove Client project
- Migrated all components, services, models, and static assets from EverettEats.Client to EverettEats (Blazor Server)
- Refactored routing, navigation, and service registration for server-side architecture
- Removed all references and files for EverettEats.Client, including project references and static assets
- Cleaned up solution and project files to eliminate asset conflicts and warnings
- Ensured build passes with only the consolidated Blazor Server project
