+++ EverettEats.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 17.14.36109.1 d17.14
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EverettEats", "EverettEats\EverettEats\EverettEats.csproj", "{135D0132-8777-4A6B-9940-60BE5CBAB4F5}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EverettEats.Client", "EverettEats\EverettEats.Client\EverettEats.Client.csproj", "{8744A439-041F-43D6-9E12-263F487E2CF0}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -17,10 +15,6 @@ Global
 		{135D0132-8777-4A6B-9940-60BE5CBAB4F5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{135D0132-8777-4A6B-9940-60BE5CBAB4F5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{135D0132-8777-4A6B-9940-60BE5CBAB4F5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8744A439-041F-43D6-9E12-263F487E2CF0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8744A439-041F-43D6-9E12-263F487E2CF0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8744A439-041F-43D6-9E12-263F487E2CF0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8744A439-041F-43D6-9E12-263F487E2CF0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
