--- EverettEats/EverettEats.Client/Components/App.razor
@@ -1,29 +1,0 @@
-@namespace EverettEats.Client.Components
-<ErrorBoundary>
-    <ChildContent>
-        <Router AppAssembly="@typeof(App).Assembly">
-            <Found Context="routeData">
-                <RouteView RouteData="@routeData" DefaultLayout="@typeof(MainLayout)" />
-                <FocusOnNavigate RouteData="@routeData" Selector="h1" />
-            </Found>
-            <NotFound>
-                <PageTitle>Not found</PageTitle>
-                <LayoutView Layout="@typeof(MainLayout)">
-                    <p role="alert">Sorry, there's nothing at this address.</p>
-                </LayoutView>
-            </NotFound>
-        </Router>
-    </ChildContent>    <ErrorContent Context="exception">
-        <div style="background: #f8fafc; border: 1px solid #1e3a8a; border-radius: 0.5rem; padding: 2rem; margin: 2rem; text-align: center; color: #1e3a8a;">
-            <h3>🚨 Application Error</h3>
-            <p>An unexpected error occurred in the application.</p>
-            <details style="margin-top: 1rem; text-align: left;">
-                <summary>Error Details</summary>
-                <pre style="background: #f5f5f5; padding: 1rem; border-radius: 0.25rem; overflow-x: auto; font-size: 0.875rem;">@exception.ToString()</pre>
-            </details>
-            <button onclick="location.reload()" style="background: #1e3a8a; color: white; border: none; padding: 0.5rem 1rem; border-radius: 0.25rem; cursor: pointer; margin-top: 1rem;">
-                Reload Page
-            </button>
-        </div>
-    </ErrorContent>
-</ErrorBoundary>
