--- EverettEats/EverettEats.Client/Components/Layout/MainLayout.razor
@@ -1,7 +1,0 @@
-@inherits LayoutComponentBase
-
-<div class="main-layout">
-    <main class="main-content">
-        @Body
-    </main>
-</div>
