--- EverettEats/EverettEats.Client/Components/LoadingSpinner.razor
@@ -1,12 +1,0 @@
-@namespace EverettEats.Client.Components
-
-<div class="loading-container">
-    <div class="loading-spinner">
-        <div class="spinner"></div>
-    </div>
-    <p class="loading-text">@Text</p>
-</div>
-
-@code {
-    [Parameter] public string Text { get; set; } = "Loading...";
-}
