--- EverettEats/EverettEats.Client/Components/_Imports.razor
@@ -1,11 +1,0 @@
-@using System.Net.Http
-@using System.Net.Http.Json
-@using Microsoft.AspNetCore.Components.Forms
-@using Microsoft.AspNetCore.Components.Routing
-@using Microsoft.AspNetCore.Components.Web
-@using static Microsoft.AspNetCore.Components.Web.RenderMode
-@using Microsoft.AspNetCore.Components.Web.Virtualization
-@using Microsoft.JSInterop
-@using EverettEats.Client
-@using EverettEats.Client.Components
-@using EverettEats.Client.Components.Layout
