--- EverettEats/EverettEats.Client/DifficultyLevel.cs
@@ -1,8 +1,0 @@
-namespace EverettEats.Client;
-
-public enum DifficultyLevel
-{
-	Easy,
-	Medium,
-	Hard
-}
