--- EverettEats/EverettEats.Client/RecipeCategory.cs
@@ -1,12 +1,0 @@
-namespace EverettEats.Client;
-
-public enum RecipeCategory
-{
-	Desserts,
-	MainDishes,
-	Appetizers,
-	Sides,
-	Beverages,
-	Breakfast,
-	Snacks
-}
