--- EverettEats/EverettEats.Client/wwwroot/data/recipes.json
@@ -1,44 +1,0 @@
-[
-  {
-    "id": 1,
-    "title": "Ultimate Fudgy Brownies",
-    "description": "Super rich, moist brownies with the perfect fudgy texture. These brownies are dense, chocolatey, and absolutely irresistible.",
-    "imageUrl": "https://images.unsplash.com/photo-1606313564200-e75d5e30476c?w=400&h=300&fit=crop",
-    "prepTime": "15 minutes",
-    "cookTime": "25-30 minutes",
-    "totalTime": "1 hour 15 minutes (includes cooling)",
-    "servings": "16 brownies",
-    "category": 0,
-    "difficulty": 0,    "dateAdded": "2025-05-17T00:00:00",
-    "slug": "ultimate-fudgy-brownies",
-    "tags": ["chocolate", "fudgy", "easy", "dessert"],
-    "ingredients": [
-      "225g (1 cup) unsalted butter",
-      "2 tablespoons vegetable oil",
-      "300g (1 1/2 cups) granulated sugar",
-      "2 whole eggs + 1 egg yolk, room temperature",
-      "1 1/2 teaspoons vanilla extract",
-      "65g (3/4 cup) unsweetened cocoa powder",
-      "63g (3/4 cup) all-purpose flour",
-      "1/4 teaspoon salt",
-      "130g (3/4 cup) chocolate, chopped (or chips)"
-    ],
-    "instructions": [
-      "Prep: Preheat oven to 175°C. Line 8x8 pan with baking paper, leaving overhang for easy removal.",
-      "Melt: Microwave butter in 30-second intervals until melted. Immediately stir in oil and sugar. Let cool 5 minutes.",
-      "Mix Wet: Beat in 2 eggs + 1 yolk one at a time, then vanilla. Mix until just combined.",
-      "Add Dry: Sift cocoa powder, flour, and salt directly over wet ingredients. Fold gently until just combined - DO NOT OVERMIX. Fold in 3/4 of the chopped chocolate.",
-      "Optional but recommended: Refrigerate batter 15 minutes for extra fudgy texture.",
-      "Bake: Pour into pan and evenly distribute the remaining 1/4 of chopped chocolate on top. Bake 25-30 minutes. The top should be set but still glossy. A toothpick should come out with moist crumbs, not clean.",
-      "Cool: Let cool in pan 30 minutes, then refrigerate for at least 1 hour before cutting for best results."
-    ],
-    "tips": [
-      "Don't Overmix: For extra fudgy brownies, don't overmix the batter once you add the flour",
-      "Perfect Doneness: The slight sinking in the middle is normal and indicates fudgy texture",
-      "Clean Cuts: Use a sharp knife and clean between cuts for neat squares",
-      "Storage: Store in an airtight container at room temperature for up to 5 days",
-      "Moisture Tip: Add a piece of bread to the container to maintain moisture if they start to dry out",
-      "Quick Cooling: Freezer method - Cool for 20 mins at room temp, then 20 mins in the freezer if you're in a hurry"
-    ]
-  }
-]
