--- EverettEats/EverettEats.Client/wwwroot/favicon.ico
@@ -1,3 +1,0 @@
-<!-- This is a placeholder favicon.ico file -->
-<!-- In a real application, you would replace this with an actual ICO image file -->
-<!-- For now, this HTML comment serves as a placeholder -->
