--- EverettEats/EverettEats.Client/wwwroot/icon-192.png
@@ -1,3 +1,0 @@
-# This is a placeholder for the 192x192 PWA icon
-# In a real application, you would replace this with an actual PNG image file
-# For now, this text file serves as a placeholder to prevent 404 errors
