--- EverettEats/EverettEats.Client/wwwroot/index.html
@@ -1,96 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Everett Eats</title>    <base href="/" />
-    <link rel="stylesheet" href="app.css" />
-    <link rel="stylesheet" href="EverettEats.Client.styles.css" />
-    
-    <!-- PWA Meta Tags -->
-    <link rel="manifest" href="/manifest.json" />
-    <meta name="theme-color" content="#b45309" />
-    <meta name="apple-mobile-web-app-capable" content="yes" />
-    <meta name="apple-mobile-web-app-status-bar-style" content="default" />
-    <meta name="apple-mobile-web-app-title" content="Everett Eats" />
-    
-    <!-- Loading styles -->
-    <style>
-        .loading-container {
-            display: flex;
-            justify-content: center;
-            align-items: center;
-            height: 100vh;
-            font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-        }
-        
-        .loading-spinner {
-            width: 40px;
-            height: 40px;
-            border: 4px solid #f3f3f3;
-            border-top: 4px solid #b45309;
-            border-radius: 50%;
-            animation: spin 1s linear infinite;
-            margin-right: 15px;
-        }
-        
-        @keyframes spin {
-            0% { transform: rotate(0deg); }
-            100% { transform: rotate(360deg); }
-        }
-    </style>
-</head>
-
-<body>
-    <div id="app">
-        <div class="loading-container">
-            <div class="loading-spinner"></div>
-            <span>Loading Everett Eats...</span>
-        </div>
-    </div>
-
-    <div id="blazor-error-ui">
-        An unhandled error has occurred.
-        <a href="" class="reload">Reload</a>
-        <a class="dismiss">🗙</a>
-    </div>
-      <script src="_framework/blazor.webassembly.js"></script>
-    <script src="js/site.js"></script>
-    
-    <!-- Enhanced Error Handling -->
-    <script>
-        window.addEventListener('unhandledrejection', function (event) {
-            console.error('Unhandled promise rejection:', event.reason);
-        });
-        
-        window.addEventListener('error', function (event) {
-            console.error('Global error:', event.error);
-        });
-        
-        // Override Blazor error handler for better logging
-        window.Blazor = window.Blazor || {};
-        window.Blazor.start = window.Blazor.start || function() {};
-        
-        // Log when Blazor starts
-        document.addEventListener('DOMContentLoaded', function() {
-            console.log('DOM loaded, waiting for Blazor...');
-        });
-    </script>
-      <!-- Service Worker Registration -->
-    <script>
-        if ('serviceWorker' in navigator) {
-            window.addEventListener('load', function() {
-                navigator.serviceWorker.register('/sw.js')
-                    .then(function(registration) {
-                        console.log('ServiceWorker registration successful');
-                    })
-                    .catch(function(err) {
-                        console.log('ServiceWorker registration failed: ', err);
-                    });
-            });
-        }
-    </script>
-</body>
-
-</html>
