--- EverettEats/EverettEats.Client/wwwroot/manifest.json
@@ -1,20 +1,0 @@
-{
-  "name": "Everett Eats",
-  "short_name": "Everett Eats",
-  "description": "Delicious homemade recipes worth sharing",
-  "start_url": "/",
-  "display": "standalone",
-  "background_color": "#fffbeb",
-  "theme_color": "#b45309",  "icons": [
-    {
-      "src": "data:image/svg+xml;charset=utf-8,%3Csvg xmlns='http://www.w3.org/2000/svg' width='192' height='192' fill='%23b45309'%3E%3Ctext y='120' font-size='120' text-anchor='middle' x='96'%3E🍽️%3C/text%3E%3C/svg%3E",
-      "sizes": "192x192",
-      "type": "image/svg+xml"
-    },
-    {
-      "src": "data:image/svg+xml;charset=utf-8,%3Csvg xmlns='http://www.w3.org/2000/svg' width='512' height='512' fill='%23b45309'%3E%3Ctext y='320' font-size='320' text-anchor='middle' x='256'%3E🍽️%3C/text%3E%3C/svg%3E",
-      "sizes": "512x512",
-      "type": "image/svg+xml"
-    }
-  ]
-}
