--- EverettEats/EverettEats.Client/wwwroot/offline.html
@@ -1,64 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Offline - Everett Eats</title>
-    <style>
-        body {
-            font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-            background-color: #fffbeb;
-            color: #1f2937;
-            text-align: center;
-            padding: 2rem;
-            margin: 0;
-        }
-        .offline-container {
-            max-width: 600px;
-            margin: 0 auto;
-            padding: 3rem 1rem;
-        }
-        .offline-icon {
-            font-size: 4rem;
-            margin-bottom: 1rem;
-        }
-        .offline-title {
-            font-size: 2rem;
-            color: #b45309;
-            margin-bottom: 1rem;
-        }
-        .offline-message {
-            font-size: 1.125rem;
-            color: #6b7280;
-            margin-bottom: 2rem;
-        }
-        .retry-button {
-            background: #b45309;
-            color: white;
-            padding: 0.75rem 2rem;
-            border: none;
-            border-radius: 0.5rem;
-            font-size: 1rem;
-            cursor: pointer;
-            text-decoration: none;
-            display: inline-block;
-        }
-        .retry-button:hover {
-            background: #92400e;
-        }
-    </style>
-</head>
-<body>
-    <div class="offline-container">
-        <div class="offline-icon">🍽️</div>
-        <h1 class="offline-title">You're Offline</h1>
-        <p class="offline-message">
-            It looks like you've lost your internet connection. 
-            Don't worry - your favorite recipes are still here waiting for you!
-        </p>
-        <button class="retry-button" onclick="window.location.reload()">
-            Try Again
-        </button>
-    </div>
-</body>
-</html>
