--- EverettEats/EverettEats.Client/wwwroot/sw.js
@@ -1,111 +1,0 @@
-const CACHE_NAME = 'everett-eats-v2';
-const urlsToCache = [
-    '/',
-    '/search',
-    '/app.css',
-    '/js/site.js',
-    '/manifest.json',
-    '/_framework/blazor.web.js',
-    '/_framework/dotnet.js',
-    '/_framework/dotnet.native.js',
-    '/_framework/dotnet.runtime.js',
-    // Cache the offline fallback page
-    '/offline.html'
-];
-
-// Install event - cache static assets
-self.addEventListener('install', event => {
-    event.waitUntil(
-        caches.open(CACHE_NAME)
-            .then(cache => {
-                console.log('Opened cache');
-                return cache.addAll(urlsToCache);
-            })
-    );
-    // Force the waiting service worker to become the active service worker
-    self.skipWaiting();
-});
-
-// Fetch event - serve from cache when offline
-self.addEventListener('fetch', event => {
-    // Only handle GET requests
-    if (event.request.method !== 'GET') {
-        return;
-    }
-
-    // For CSS and JS files, always try network first for fresh updates
-    if (event.request.url.includes('.css') || event.request.url.includes('.js')) {
-        event.respondWith(
-            fetch(event.request)
-                .then(fetchResponse => {
-                    // Cache the fresh response
-                    if (fetchResponse && fetchResponse.status === 200) {
-                        const responseToCache = fetchResponse.clone();
-                        caches.open(CACHE_NAME)
-                            .then(cache => {
-                                cache.put(event.request, responseToCache);
-                            });
-                    }
-                    return fetchResponse;
-                })
-                .catch(() => {
-                    // Fallback to cache if network fails
-                    return caches.match(event.request);
-                })
-        );
-        return;
-    }
-
-    event.respondWith(
-        caches.match(event.request)
-            .then(response => {
-                // Return cached version if available
-                if (response) {
-                    return response;
-                }
-
-                // Try to fetch from network
-                return fetch(event.request)
-                    .then(fetchResponse => {
-                        // Check if we received a valid response
-                        if (!fetchResponse || fetchResponse.status !== 200 || fetchResponse.type !== 'basic') {
-                            return fetchResponse;
-                        }
-
-                        // Clone the response as it can only be consumed once
-                        const responseToCache = fetchResponse.clone();
-
-                        caches.open(CACHE_NAME)
-                            .then(cache => {
-                                cache.put(event.request, responseToCache);
-                            });
-
-                        return fetchResponse;
-                    })
-                    .catch(() => {
-                        // Network failed, try to return cached page or offline fallback
-                        if (event.request.destination === 'document') {
-                            return caches.match('/offline.html');
-                        }
-                    });
-            })
-    );
-});
-
-// Activate event - clean up old caches
-self.addEventListener('activate', event => {
-    event.waitUntil(
-        caches.keys().then(cacheNames => {
-            return Promise.all(
-                cacheNames.map(cacheName => {
-                    if (cacheName !== CACHE_NAME) {
-                        console.log('Deleting old cache:', cacheName);
-                        return caches.delete(cacheName);
-                    }
-                })
-            );
-        })
-    );
-    // Ensure the new service worker takes control immediately
-    self.clients.claim();
-});
