+++ EverettEats/EverettEats/Components/Layout/MainLayout.razor
@@ -1,0 +1,1 @@
+// ...existing code from EverettEats.Client/Components/Layout/MainLayout.razor...
