+++ EverettEats/EverettEats/Components/Pages/Error.razor
@@ -1,0 +1,1 @@
+// ...existing code from EverettEats.Client/Components/Pages/Error.razor...
