+++ EverettEats/EverettEats/Components/Pages/Search.razor
@@ -1,0 +1,1 @@
+// ...existing code from EverettEats.Client/Components/Pages/Search.razor...
