+++ EverettEats/EverettEats/Components/Routes.razor
@@ -1,6 +1,6 @@
-﻿<Router AppAssembly="typeof(EverettEats.Client._Imports).Assembly">
+﻿<Router AppAssembly="typeof(Program).Assembly">
     <Found Context="routeData">
-        <RouteView RouteData="routeData" DefaultLayout="typeof(EverettEats.Client.Components.Layout.MainLayout)" />
+        <RouteView RouteData="routeData" DefaultLayout="typeof(EverettEats.Components.Layout.MainLayout)" />
         <FocusOnNavigate RouteData="routeData" Selector="h1" />
     </Found>
 </Router>
