+++ EverettEats/EverettEats/Components/_Imports.razor
@@ -7,5 +7,5 @@
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.JSInterop
 @using EverettEats
-@using EverettEats.Client
+// Removed EverettEats.Client reference after consolidation
 @using EverettEats.Components
