+++ EverettEats/EverettEats/Models/DifficultyLevel.cs
@@ -1,0 +1,9 @@
+namespace EverettEats.Models
+{
+	public enum DifficultyLevel
+	{
+		Easy,
+		Medium,
+		Hard
+	}
+}
