+++ EverettEats/EverettEats/Models/RecipeCategory.cs
@@ -1,0 +1,13 @@
+namespace EverettEats.Models
+{
+	public enum RecipeCategory
+	{
+		Desserts,
+		MainDishes,
+		Appetizers,
+		Sides,
+		Beverages,
+		Breakfast,
+		Snacks
+	}
+}
