+++ EverettEats/EverettEats/wwwroot/app.css
@@ -1,0 +1,1 @@
+// ...existing code from EverettEats.Client/wwwroot/app.css...
