+++ EverettEats/EverettEats/wwwroot/data/recipes.json
@@ -1,0 +1,1 @@
+// ...existing code from EverettEats.Client/wwwroot/data/recipes.json...
