+++ EverettEats/EverettEats/wwwroot/favicon.ico
@@ -1,0 +1,1 @@
+// ...binary content from EverettEats.Client/wwwroot/favicon.ico...
