+++ EverettEats/EverettEats/wwwroot/icon-512.png
@@ -1,0 +1,1 @@
+// ...binary content from EverettEats.Client/wwwroot/icon-512.png...
