+++ EverettEats/EverettEats/wwwroot/index.html
@@ -1,0 +1,1 @@
+// ...existing code from EverettEats.Client/wwwroot/index.html...
