+++ EverettEats/EverettEats/wwwroot/js/site.js
@@ -1,0 +1,1 @@
+// ...existing code from EverettEats.Client/wwwroot/js/site.js...
