+++ EverettEats/EverettEats/wwwroot/manifest.json
@@ -1,0 +1,1 @@
+// ...existing code from EverettEats.Client/wwwroot/manifest.json...
