+++ EverettEats/EverettEats/wwwroot/offline.html
@@ -1,0 +1,1 @@
+// ...existing code from EverettEats.Client/wwwroot/offline.html...
