+++ EverettEats/EverettEats/wwwroot/sw.js
@@ -1,0 +1,1 @@
+// ...existing code from EverettEats.Client/wwwroot/sw.js...
